feat: use glbinding instead of glad
Because glad generates code it makes the build process hard to manage
without vendoring the generated code, which makes different configs hard
to manage. glbinding provides everything that glad provides, plus
enhanced typesafety and robust debugging support.

This did require reorganizing the structure of the code, since glbinding
is not compatible with libraries that import gl.h (namely glfw), so the
window management code has to be kept separate from the renderering
code. This is an architectural improvement anyway though.

Finally, I've removed conan since glbinding isn't available on conan
while all of our libraries seem to be available on vcpkg. It just seems
like vcpkg is gaining more traction than conan and doesn't clutter the build
environment or config as a bonus.