feat: tile-based rendering abstraction
BREAKING CHANGE: This commit adds a huge number of changes, most notibly
in that it adds an abstraction layer to Renderer to efficiently render
a grid of tiles w/ separate routines for updating and painting them. The
Layer abstraction in particular helps codify the interface for these
tile layers.

This also adds unit tests of Layer and the pprint library to faciliate unit
testing.